SudokuSolverAlgo
In this part, the main algorithms of the application are implemented. In fact, these algorithms are designed to solve Sudoku problems that will be entered in the future by the users of the application. There are several algorithms, we chose one of them as the main one after efficiency tests. Of course, it is possible to add algorithms freely because this package is added as a JAR on the server side, which allows for proper protection and division of work and the ability to change the algorithmic part of the project without any changes on the server side.